refactor: change state colors from student view
